--- app/Filament/Resources/KegiatanMagangResource/Pages/CreateKegiatanMagang.php
@@ -1,12 +1,0 @@
-<?php
-
-namespace App\Filament\Resources\KegiatanMagangResource\Pages;
-
-use App\Filament\Resources\KegiatanMagangResource;
-use Filament\Actions;
-use Filament\Resources\Pages\CreateRecord;
-
-class CreateKegiatanMagang extends CreateRecord
-{
-    protected static string $resource = KegiatanMagangResource::class;
-}
